Add test suite
* updated export, file format is changed
* added utilities in tests and kite module
* set random seed

+++ kite/export_kite.py
@@ -1663,7 +1663,7 @@ def config_system(lattice, config, calculation, modification=None, **kwargs):
             raise SystemExit('Only a single function request of each type is currently allowed. Please use another '
                              'configuration file for the same functionality.')
         grpc_p.create_dataset('NumMoments', data=moments, dtype=np.int32)
-        grpc_p.create_dataset('Energy', data=(np.asarray(energy) - config.energy_shift) / config.energy_scale, dtype=np.float32)
+        grpc_p.create_dataset('Energy', data=(np.asarray(energy) - config.energy_shift) / config.energy_scale, dtype=np.float64)
         if len_sub != len_pos:
             grpc_p.create_dataset('Orbitals', data=np.tile(np.asarray(orbitals),len_pos).reshape(-1), dtype=np.int32)
             grpc_p.create_dataset('FixPosition', data=np.repeat(np.asarray(fixed_positions),len_sub), dtype=np.int32)
@@ -1695,7 +1695,7 @@ def config_system(lattice, config, calculation, modification=None, **kwargs):
             raise SystemExit('Only a single function request of each type is currently allowed. Please use another '
                              'configuration file for the same functionality.')
         grpc_p.create_dataset('NumMoments', data=moments, dtype=np.int32)
-        grpc_p.create_dataset('k_vector', data=np.asmatrix(np.asarray(k_vector_rel)), dtype=np.float32)
+        grpc_p.create_dataset('k_vector', data=np.asmatrix(np.asarray(k_vector_rel)), dtype=np.float64)
         grpc_p.create_dataset('NumDisorder', data=dis, dtype=np.int32)
         grpc_p.create_dataset('OrbitalWeights', data=np.asmatrix(np.asarray(spinor)))
 
@@ -1721,12 +1721,12 @@ def config_system(lattice, config, calculation, modification=None, **kwargs):
         grpc_p.create_dataset('NumMoments', data=num_moments, dtype=np.int32)
         grpc_p.create_dataset('NumPoints', data=num_points, dtype=np.int32)
         grpc_p.create_dataset('NumDisorder', data=num_disorder, dtype=np.int32)
-        grpc_p.create_dataset('mean_value', data=mean_value, dtype=np.float32)
-        grpc_p.create_dataset('ProbingPoint', data=np.asmatrix(np.asarray(probing_points)).astype(np.float32))
+        grpc_p.create_dataset('mean_value', data=mean_value, dtype=np.float64)
+        grpc_p.create_dataset('ProbingPoint', data=np.asmatrix(np.asarray(probing_points)), dtype=np.float64)
         grpc_p.create_dataset('width', data=width, dtype=np.float)
         grpc_p.create_dataset('spinor', data=np.asmatrix(np.asarray(spinor)).astype(config.type))
-        grpc_p.create_dataset('k_vector', data=np.asmatrix(np.asarray(k_vector)), dtype=np.float32)
-        grpc_p.create_dataset('timestep', data=timestep, dtype=np.float32)
+        grpc_p.create_dataset('k_vector', data=np.asmatrix(np.asarray(k_vector)), dtype=np.float64)
+        grpc_p.create_dataset('timestep', data=timestep, dtype=np.float64)
 
     if calculation.get_conductivity_dc:
         grpc_p = grpc.create_group('conductivity_dc')
